+++ src/components/CoreConcepts.jsx
@@ -1,11 +1,23 @@
 import "./CoreConcepts.css"
+import { CORE_CONCEPTS } from "../data/data-with-examples"
 
-export default function CoreConcepts ({image, title, description}) {
+function CoreConcepts ({image, title, description}) {
     return (
         <li>
             <img src={image} alt={title} />
             <h3> {title} </h3>
             <p> {description} </p>
         </li>
+    )
+}
+
+export default function CoreConceptsBloc (props) {
+    return(
+        <section id="core-concepts">
+            <h2>Core Concepts</h2>
+            <ul>
+                {CORE_CONCEPTS.map((conceptItem) => <CoreConcepts key={conceptItem.title} {...conceptItem} />)}
+            </ul>
+        </section>
     )
 }
